Renamed api.js to order.js and updated server.js
